+++ src/plonemeeting/portal/core/browser/static/js/core.js
@@ -1,4 +1,15 @@
+function setUpEnvironmentLabel() {
+    var hostname = document.location.hostname;
+    if (hostname === 'localhost' || hostname === '0.0.0.0') {
+        $("body").append("<span class='environment-label environment-dev'>DEV</span>");
+    }
+    else if (hostname.includes("staging")) {
+        $("body").append("<span class='environment-label environment-test'>TEST</span>");
+    }
+}
+
 jQuery(document).ready(function ($) {
+  setUpEnvironmentLabel();
   if (window.Faceted) {
     jQuery(Faceted.Events).bind(Faceted.Events.AJAX_QUERY_SUCCESS, function () {
       $(".toggle-link").click(function (e) {